refactor: Move `_replace_preprocessing_names` to utils.py
